src/faceInstanceApp/requirements_AMD64.txt
flask
dapr
flask-dapr
gunicorn
Pillow
requests
numpy
opencv-contrib-python==4.5.3.56
azure-cognitiveservices-vision-face==0.5.0